try catch around program.run(), no more problems!

+++ Client/AuctionPimp.cs
@@ -10,7 +10,7 @@ namespace Client
     {
         public int Id { get; set; }
         public int StartPrice { get; set; }
-        public int ExpDate { get; set; }
+        public long ExpDate { get; set; }
         public Item BidItem { get; set; }
         public int CurrentBid { get; set; }
         public override string ToString()
